+++ salvo/src/lib/docker_management/docker_image.py
